requirements.txt
alabaster==1.0.0
anaconda-anon-usage==0.4.4
anndata==0.11.4
annotated-types==0.7.0
anthropic==0.57.1
anyio==4.9.0
archspec==0.2.3
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
array-api-compat==1.12.0
arrow==1.3.0
asciitree==0.3.3
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.3
bioframe==0.7.2
black==25.1.0
bleach==6.2.0
blinker==1.9.0
boltons==23.0.0
Brotli==1.0.9
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.8
comm==0.2.2
conda==25.7.0
conda-build==25.1.2
conda-content-trust==0.2.0
conda_index==0.5.0
conda-libmamba-solver==24.11.1
conda-package-handling==2.3.0
conda_package_streaming==0.10.0
contourpy==1.3.1
cooler==0.10.3
cooltools==0.7.1
cryptography==43.0.0
cycler==0.12.1
Cython==3.0.12
cytoolz==1.0.1
dadi==2.4.3
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
demes==0.2.3
demesdraw==0.4.0
dill==0.3.9
distro==1.9.0
docutils==0.21.2
executing==2.2.0
faiss-cpu==1.11.0
fastjsonschema==2.21.1
filelock==3.18.0
filetype==1.2.0
flake8==7.3.0
Flask==3.1.0
fonttools==4.56.0
fqdn==1.5.1
frozendict==2.4.2
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.73.1
h11==0.16.0
h5py==3.13.0
httpcore==1.0.9
httpx==0.28.1
idna==3.7
igraph==0.11.9
imageio==2.37.0
imagesize==1.4.1
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.4.2
json5==0.12.0
jsonpatch==1.33
jsonpointer==2.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter_client==8.6.3
jupyter-console==6.6.3
jupyter_core==5.8.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
langchain==0.3.26
langchain-anthropic==0.3.17
langchain-core==0.3.68
langchain-google-genai==2.1.6
langchain-openai==0.3.27
langchain-text-splitters==0.3.8
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.8
langsmith==0.4.4
lazy_loader==0.4
legacy-api-wrap==1.4.1
libarchive-c==5.2
libmambapy==2.0.4
lief==0.14.1
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mccabe==0.7.0
menuinst==2.1.2
mistune==3.1.3
more-itertools==10.6.0
multiprocess==0.70.17
mypy_extensions==1.1.0
natsort==8.4.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nlopt==2.9.1
notebook==7.4.4
notebook_shim==0.2.4
numba==0.61.0
numpy==1.26.4
openai==1.93.1
orjson==3.10.18
ormsgpack==1.10.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
patsy==1.0.1
pexpect==4.9.0
pillow==11.1.0
pip==24.2
pkginfo==1.12.1.2
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
pluggy==1.6.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycosat==0.6.6
pycparser==2.21
pydantic==2.11.7
pydantic_core==2.33.2
pyfaidx==0.8.1.3
pyflakes==3.4.0
Pygments==2.19.1
pynndescent==0.5.13
pyparsing==3.2.3
PySocks==1.7.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-igraph==0.11.9
python-json-logger==3.3.0
pytz==2025.1
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
roman-numerals-py==3.1.0
rpds-py==0.23.1
rsa==4.9.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scanpy==1.11.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
Send2Trash==1.8.3
session-info2==0.1.2
setuptools==75.1.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==8.2.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.41
stack-data==0.6.3
statsmodels==0.14.4
tenacity==9.1.2
terminado==0.18.1
texttable==1.7.0
threadpoolctl==3.6.0
tifffile==2025.3.13
tiktoken==0.9.0
tinycss2==1.4.0
toolz==1.0.0
tornado==6.5.1
tqdm==4.66.5
traitlets==5.14.3
truststore==0.8.0
types-python-dateutil==2.9.0.20250516
typing_extensions==4.12.2
typing-inspection==0.4.1
tzdata==2025.2
umap-learn==0.5.9.post2
uri-template==1.3.0
urllib3==2.2.3
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.44.0
widgetsnbextension==4.0.14
xxhash==3.5.0
zipp==3.21.0
zstandard==0.23.0
